+++ R.NET/ComplexVector.cs
@@ -66,7 +66,7 @@ namespace RDotNet
         /// <returns>The element at the specified index.</returns>
         protected override Complex GetValueAltRep(int index)
         {
-            var data = GetFunction<COMPLEX_ELT>()(this.DangerousGetHandle(), (ulong)index);
+            var data = GetFunction<COMPLEX_ELT>()(this.DangerousGetHandle(), (IntPtr)index);
             return new Complex(data.r, data.i);
         }
 
@@ -92,7 +92,7 @@ namespace RDotNet
         /// <param name="value">The value to set</param>
         protected override void SetValueAltRep(int index, Complex value)
         {
-            GetFunction<SET_COMPLEX_ELT>()(this.DangerousGetHandle(), (ulong)index,
+            GetFunction<SET_COMPLEX_ELT>()(this.DangerousGetHandle(), (IntPtr)index,
                         RTypesUtil.SerializeComplexToRComplex(value));
         }
