+++ R.NET/IntegerVector.cs
@@ -76,7 +76,7 @@ namespace RDotNet
         /// <returns>The element at the specified index.</returns>
         protected override int GetValueAltRep(int index)
         {
-            return GetFunction<INTEGER_ELT>()(this.DangerousGetHandle(), (ulong)index);
+            return GetFunction<INTEGER_ELT>()(this.DangerousGetHandle(), (IntPtr)index);
         }
 
         /// <summary>
@@ -99,7 +99,7 @@ namespace RDotNet
         /// <param name="value">The value to set</param>
         protected override void SetValueAltRep(int index, int value)
         {
-            GetFunction<SET_INTEGER_ELT>()(this.DangerousGetHandle(), (ulong)index, value);
+            GetFunction<SET_INTEGER_ELT>()(this.DangerousGetHandle(), (IntPtr)index, value);
         }
 
         /// <summary>
@@ -122,7 +122,7 @@ namespace RDotNet
         {
             // by inference from `static SEXP compact_intseq_Duplicate(SEXP x, Rboolean deep)`  in altrep.c
             var res = new int[this.Length];
-            GetFunction<INTEGER_GET_REGION>()(this.DangerousGetHandle(), (ulong)0, (ulong)this.Length, res);
+            GetFunction<INTEGER_GET_REGION>()(this.DangerousGetHandle(), (IntPtr)0, (IntPtr)this.Length, res);
             return res;
         }
